Chore: 프로젝트 폴더 구조 변경, App.jsx 라우팅 구조 변경
- 최상단 폴더 제거
- App.jsx: Layout 컴포넌트 연결
