--- mung-ng-nyang/src/components/Layout.jsx
@@ -1,10 +1,0 @@
-
-const Layout = () => {
-    return (
-        <>
-
-        </>
-    )
-}
-
-export default Layout;
