--- mung-ng-nyang/src/pages/myPage/MyInfo.jsx
@@ -1,11 +1,0 @@
-const MyInfo = () => {
-  
-
-    return (
-      <>
-
-      </>
-    )
-  }
-  
-  export default MyInfo
