--- mung-ng-nyang/src/pages/myPage/MyPetInfo.jsx
@@ -1,11 +1,0 @@
-const MyPetInfo = () => {
-  
-
-    return (
-      <>
-
-      </>
-    )
-  }
-  
-  export default MyPetInfo
