+++ src/components/Layout.jsx
@@ -1,0 +1,11 @@
+import { Outlet } from 'react-router-dom';
+
+const Layout = () => {
+  return (
+    <>
+      <Outlet />
+    </>
+  );
+};
+
+export default Layout;
